feat(kscan): Move temp state into debouncer
Changed matrix scanning so instead of storing a temporary read state
for each key while scanning, then debouncing the matrix all at once, it
debounces each key as it reads it and tracks whether the key changed in
the debounce state, then checks which keys changed all at once later.

This avoids needing parallel arrays, and it gives us 14 bits for the
counter instead of 7 with no increase in memory size.